+++ test/integ/pandas/test_pandas_tools.py
@@ -14,9 +14,9 @@ import pytest
 
 from snowflake.connector import DictCursor
 from snowflake.connector.errors import ProgrammingError
+from snowflake.connector.util_text import random_string
 
 from ...lazy_var import LazyVar
-from ...randomize import random_string
 
 try:
     from snowflake.connector.options import pandas
