+++ test/integ/test_bindings.py
@@ -20,8 +20,7 @@ import pytz
 
 from snowflake.connector.converter import convert_datetime_to_epoch
 from snowflake.connector.errors import ForbiddenError, ProgrammingError
-
-from ..randomize import random_string
+from snowflake.connector.util_text import random_string
 
 tempfile.gettempdir()
