+++ test/integ/test_cursor.py
@@ -51,8 +51,7 @@ from snowflake.connector.errorcode import (
 )
 from snowflake.connector.sqlstate import SQLSTATE_FEATURE_NOT_SUPPORTED
 from snowflake.connector.telemetry import TelemetryField
-
-from ..randomize import random_string
+from snowflake.connector.util_text import random_string
 
 try:
     from snowflake.connector.constants import (
