+++ test/integ/test_put_get.py
@@ -17,10 +17,10 @@ from unittest import mock
 import pytest
 
 from snowflake.connector import OperationalError
+from snowflake.connector.util_text import random_string
 
 from ..generate_test_files import generate_k_lines_of_n_files
 from ..integ_helpers import put
-from ..randomize import random_string
 
 if TYPE_CHECKING:
     from snowflake.connector import SnowflakeConnection
