+++ test/integ/test_put_get_medium.py
@@ -23,10 +23,10 @@ from snowflake.connector.file_transfer_agent import (
     SnowflakeProgressPercentage,
     SnowflakeS3ProgressPercentage,
 )
+from snowflake.connector.util_text import random_string
 
 from ..generate_test_files import generate_k_lines_of_n_files
 from ..integ_helpers import put
-from ..randomize import random_string
 
 if TYPE_CHECKING:
     from snowflake.connector import SnowflakeConnection
