+++ test/integ/test_put_get_user_stage.py
@@ -15,10 +15,10 @@ from unittest.mock import patch
 import pytest
 
 from snowflake.connector.cursor import SnowflakeCursor
+from snowflake.connector.util_text import random_string
 
 from ..generate_test_files import generate_k_lines_of_n_files
 from ..integ_helpers import put
-from ..randomize import random_string
 
 
 @pytest.mark.aws
