+++ test/integ/test_put_get_with_azure_token.py
@@ -19,10 +19,10 @@ from snowflake.connector.file_transfer_agent import (
     SnowflakeAzureProgressPercentage,
     SnowflakeProgressPercentage,
 )
+from snowflake.connector.util_text import random_string
 
 from ..generate_test_files import generate_k_lines_of_n_files
 from ..integ_helpers import put
-from ..randomize import random_string
 
 logger = getLogger(__name__)
