+++ test/integ/test_put_get_with_gcp_account.py
@@ -33,9 +33,10 @@ except ImportError:
 
     SnowflakeGCSRestClient = None
 
+from snowflake.connector.util_text import random_string
+
 from ..generate_test_files import generate_k_lines_of_n_files
 from ..integ_helpers import put
-from ..randomize import random_string
 
 # We need these for our OldDriver tests. We run most up to date tests with the oldest supported driver version
 try:
