+++ test/unit/test_gcs_client.py
@@ -14,8 +14,7 @@ import pytest
 
 from snowflake.connector import SnowflakeConnection
 from snowflake.connector.constants import SHA256_DIGEST, ResultStatus
-
-from ..randomize import random_string
+from snowflake.connector.util_text import random_string
 
 try:
     from snowflake.connector.errors import RequestExceedMaxRetryError
